agentV0/requirements.txt
browser-use>=0.11.0
playwright>=1.49.0
supabase>=2.11.0
python-dotenv>=1.0.0
openai>=1.60.0
anthropic>=0.42.0
